Change proxy and refactoring
