+++ app/containers/ToDos/constants.js
@@ -3,7 +3,7 @@
  * Call TODO constants
  *
  */
-export const CLICK = 'TO_DO_CLICK';
+export const CALL_TODO_LIST = 'TO_DO_CLICK';
 export const CALL_TODO_SUCCESS = 'TO_DO_SUCCESS';
 export const CALL_TODO_REJECTED = 'TO_DO_ERROR';
 export const DELETE_TODO = 'TO_DO_DELETE';
