+++ app/services/getToDoList.js
@@ -1,0 +1,9 @@
+import proxy from 'utils/networkProxy';
+
+// axios instance setting
+export default function getToDoList() {
+  return proxy({
+    method: 'get',
+    url: 'toDos',
+  });
+}
