+++ app/utils/networkProxy.js
@@ -5,10 +5,4 @@ import axios from 'axios';
 const proxy = axios.create();
 proxy.defaults.baseURL = process.env.URL_TO_DO_API;
 
-// axios instance setting
-export default function getToDoList() {
-  return proxy({
-    method: 'get',
-    url: 'toDos',
-  });
-}
+export default proxy;
